backend/requirements.txt
Flask
flask-security-too
flask-cors
Flask-SQLAlchemy
Flask-Caching
redis
celery
